autoconf: improve appliance of verbosity settings
Autoconf adds consideration for the 'V' environment variable to generated
Makefiles in order to determine make verbosity. This variable may interfere
with the '-s' flag as well as with verbosity settings made during the call to
configure. With this commit, 'V' is getting set in accordance with the presence
of the goa --verbose flag when calling make in the autoconf case.

Fixes #19